+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import Header from "@/components/Header";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -17,8 +18,6 @@ export const metadata: Metadata = {
   description: "私の技術ポートフォリオサイトです",
 };
 
-import Link from "next/link";
-
 export default function RootLayout({
   children,
 }: Readonly<{
@@ -29,37 +28,7 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        {/* Navigation */}
-        <nav className="bg-white shadow-lg">
-          <div className="container mx-auto px-4">
-            <div className="flex justify-between items-center h-16">
-              <Link href="/" className="text-xl font-bold text-gray-900">
-                My technical history
-              </Link>
-              <div className="hidden md:flex space-x-8">
-                <Link
-                  href="/"
-                  className="text-gray-600 hover:text-blue-600 transition-colors"
-                >
-                  ホーム
-                </Link>
-                <Link
-                  href="/about"
-                  className="text-gray-600 hover:text-blue-600 transition-colors"
-                >
-                  自己紹介
-                </Link>
-                <Link
-                  href="/projects"
-                  className="text-gray-600 hover:text-blue-600 transition-colors"
-                >
-                  プロジェクト
-                </Link>
-              </div>
-            </div>
-          </div>
-        </nav>
-
+        <Header />
         {children}
       </body>
     </html>
