+++ src/components/Footer.tsx
@@ -1,0 +1,11 @@
+export default function Footer() {
+  return (
+    <footer className="bg-gray-50 py-4">
+      <div className="container mx-auto px-4 text-center">
+        <p className="text-gray-500 text-sm">
+          © 2024 My technical history. All rights reserved.
+        </p>
+      </div>
+    </footer>
+  );
+}
