+++ src/components/Header.tsx
@@ -1,0 +1,35 @@
+import Link from "next/link";
+
+export default function Header() {
+  return (
+    <nav className="bg-white shadow-lg">
+      <div className="container mx-auto px-4">
+        <div className="flex justify-between items-center h-16">
+          <Link href="/" className="text-xl font-bold text-gray-900">
+            My technical history
+          </Link>
+          <div className="hidden md:flex space-x-8">
+            <Link
+              href="/"
+              className="text-gray-600 hover:text-blue-600 transition-colors"
+            >
+              ホーム
+            </Link>
+            <Link
+              href="/about"
+              className="text-gray-600 hover:text-blue-600 transition-colors"
+            >
+              自己紹介
+            </Link>
+            <Link
+              href="/projects"
+              className="text-gray-600 hover:text-blue-600 transition-colors"
+            >
+              プロジェクト
+            </Link>
+          </div>
+        </div>
+      </div>
+    </nav>
+  );
+}
