+++ src/const/projects.ts
@@ -1,0 +1,26 @@
+export const projects = [
+  {
+    id: "co2-calculator",
+    title: "CO₂排出量計算アプリ",
+    description:
+      "家庭や生活スタイルから、月間のCO₂排出量を簡単に計算できるWebアプリ。電気・ガス・水道、交通手段、食生活の影響を総合的に計算します。",
+    technologies: ["Next.js 15", "TypeScript", "Tailwind CSS", "Framer Motion"],
+    category: "環境技術",
+    image: "🌱",
+    status: "完了",
+    linkColor: "text-[#2C5D47] hover:text-green-800",
+    bgColor: "bg-green-100",
+  },
+  {
+    id: "dinner-record",
+    title: "Dinner Record",
+    description:
+      "日々の夕食を簡単に記録・管理できるWebアプリケーション。シンプルなUIで料理名やメモを入力し、過去の食事を一覧表示・編集・削除できます。",
+    technologies: ["Next.js", "TypeScript", "Tailwind CSS", "Firebase"],
+    category: "Webアプリ",
+    image: "🍽",
+    status: "完了",
+    linkColor: "text-orange-600 hover:text-orange-800",
+    bgColor: "bg-orange-100",
+  },
+];
